♻️ string 날짜를 object로 바꿔주는 함수를 helperFunction으로 분리
