+++ client/src/pages/Booking/store/MonthlyReservationStore.tsx
@@ -1,37 +1,8 @@
 import { configureStore, createSlice, PayloadAction } from '@reduxjs/toolkit';
-import { StartEndDateProps } from '../model/IStartEndDateProps';
+import { IMonthlyReservation } from '../model/IMonthlyReservation';
+import { turnStringArrIntoDateObjectArr } from '../../../common/utils/helperFunctions/turnStringArrIntoDateObjectArr';
 
-const turnStringArrIntoDateObjectArr = (
-  date: { startDate: string; endDate: string }[],
-) => {
-  const dateObjectArr = date.map(
-    (dateObj: { startDate: string; endDate: string }) => {
-      const startDate = {
-        year: Number(dateObj.startDate.slice(0, 4)),
-        month: Number(dateObj.startDate.slice(5, 7)),
-        date: Number(dateObj.startDate.slice(8, 10)),
-      };
-      const endDate = {
-        year: Number(dateObj.endDate.slice(0, 4)),
-        month: Number(dateObj.endDate.slice(5, 7)),
-        date: Number(dateObj.endDate.slice(8, 10)),
-      };
-      return { startDate, endDate };
-    },
-  );
-  return dateObjectArr;
-};
-
-interface MonthlyReservationProps {
-  productTitle: string;
-  baseFee: number;
-  feePerDay: number;
-  minimumRentalPeriod: number;
-  reservationsDate1: StartEndDateProps[];
-  reservationsDate2: StartEndDateProps[];
-}
-
-const initialMonthlyReservationState: MonthlyReservationProps = {
+const initialMonthlyReservationState: IMonthlyReservation = {
   productTitle: '제목',
   baseFee: 100,
   feePerDay: 10,
@@ -96,7 +67,7 @@ export const monthlyReservation = createSlice({
   reducers: {
     setMonthlyReservation: (
       state,
-      action: PayloadAction<MonthlyReservationProps>,
+      action: PayloadAction<IMonthlyReservation>,
     ) => {
       state.productTitle = action.payload.productTitle;
       state.baseFee = action.payload.baseFee;
